requirements.txt
absl-py==0.7.1
astor==0.7.1
certifi==2019.3.9
cycler==0.10.0
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.3
mkl-fft==1.0.10
mkl-random==1.0.2
mock==2.0.0
numpy==1.16.2
olefile==0.46
pandas==0.24.2
patsy==0.5.1
pbr==5.1.3
Pillow==5.4.1
protobuf==3.7.0
pydot==1.4.1
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.1
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
six==1.12.0
statsmodels==0.9.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
tornado==6.0.1
Werkzeug==0.14.1
